Added React-Route
Added routing and starting to "Componentize".

Added a wrapper to the components.

Changed "StarterPage" component name to "Overview" to be more intuitive.

N.B. The rng present in the components it's to test the Routing page regen
I.E. I never used React Routing so i don't knew the page was regenerated after re-entering without Reloading the page
E.G. Changing to another page and returning to it